requirements.txt
sanic~=23.6.0
sanic-ext~=23.6.0
sqlalchemy~=2.0.20
python-dotenv==1.0.0
email-validator==2.0.0.post2
contextvars==2.4
WTForms~=3.0.1
alembic~=1.11.3
PyJWT~=2.8.0
aiofiles~=23.2.1
yookassa~=2.4.0
bcrypt==4.0.1